+++ provision/ubuntu.sh
@@ -6,6 +6,7 @@ set -e -u
 
 mt6-include linux
 
+export JAIL_DEVFS_RULESET="$JAIL_DEVFS_RULESET_LINUX"
 export JAIL_START_EXTRA="allow.mount
 		allow.mount.devfs
 		allow.mount.fdescfs
@@ -18,11 +19,6 @@ export JAIL_START_EXTRA="allow.mount
 export JAIL_CONF_EXTRA='
 		allow.raw_sockets;'
 export JAIL_FSTAB="
-devfs     $ZFS_JAIL_MNT/ubuntu/compat/linux/dev     devfs     rw,late  0 0
-tmpfs     $ZFS_JAIL_MNT/ubuntu/compat/linux/dev/shm tmpfs     rw,late,size=1g,mode=1777  0 0
-fdescfs   $ZFS_JAIL_MNT/ubuntu/compat/linux/dev/fd  fdescfs   rw,late,linrdlnk 0 0
-linprocfs $ZFS_JAIL_MNT/ubuntu/compat/linux/proc    linprocfs rw,late  0 0
-linsysfs  $ZFS_JAIL_MNT/ubuntu/compat/linux/sys     linsysfs  rw,late  0 0
 #/tmp      $ZFS_JAIL_MNT/ubuntu/compat/linux/tmp     nullfs    rw,late  0 0
 #/home     $ZFS_JAIL_MNT/ubuntu/compat/linux/home    nullfs    rw,late  0 0"
 
@@ -38,6 +34,7 @@ for _fs in dev proc sys tmp home; do
 	mkdir -p "$STAGE_MNT/compat/linux/$_fs"
 done
 chmod 777 "$STAGE_MNT/compat/linux/tmp"
+assure_devfs_linux_ruleset
 start_staged_jail ubuntu
 install_ubuntu
 promote_staged_jail ubuntu